fix type error (double vs float)

+++ lib/datasets/aflw.py
@@ -40,8 +40,8 @@ class AFLW(data.Dataset):
         self.rot_factor = cfg.DATASET.ROT_FACTOR
         self.label_type = cfg.MODEL.TARGET_TYPE
         self.flip = cfg.DATASET.FLIP
-        self.mean = np.array([0.485, 0.456, 0.406])
-        self.std = np.array([0.229, 0.224, 0.225])
+        self.mean = np.array([0.485, 0.456, 0.406], dtype=np.float32)
+        self.std = np.array([0.229, 0.224, 0.225], dtype=np.float32)
         # load annotations
         self.landmarks_frame = pd.read_csv(self.csv_file)
 
@@ -65,7 +65,7 @@ class AFLW(data.Dataset):
 
         scale *= 1.25
         nparts = pts.shape[0]
-        img = np.array(Image.open(image_path).convert('RGB'))
+        img = np.array(Image.open(image_path).convert('RGB'), dtype=np.float32)
 
         # transform !
         # img = self.transform(img)
